Add underscore to internal modules, move tools for PC fitting to their own module
