+++ src/kikuchipy/io/plugins/kikuchipy_h5ebsd/_api.py
@@ -32,7 +32,7 @@ from orix.io.plugins.orix_hdf5 import crystalmap2dict, dict2crystalmap
 from rsciio.hspy._api import overwrite_dataset
 
 from kikuchipy import __version__ as kikuchipy_version
-from kikuchipy.detectors.ebsd_detector import EBSDDetector
+from kikuchipy.detectors._ebsd_detector import EBSDDetector
 from kikuchipy.io._util import _get_input_variable
 from kikuchipy.io.plugins._h5ebsd import H5EBSDReader, _dict2hdf5group, _hdf5group2dict
 from kikuchipy.signals.util._crystal_map import _xmap_is_compatible_with_signal
