+++ src/kikuchipy/io/plugins/nordif_calibration_patterns/_api.py
@@ -1,4 +1,5 @@
-# Copyright 2019-2024 The kikuchipy developers
+#
+# Copyright 2019-2026 the kikuchipy developers
 #
 # This file is part of kikuchipy.
 #
@@ -14,6 +15,7 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
+#
 
 """Reader of EBSD calibration patterns from NORDIF files."""
 
@@ -24,7 +26,7 @@ import warnings
 from matplotlib.pyplot import imread
 import numpy as np
 
-from kikuchipy.detectors.ebsd_detector import EBSDDetector
+from kikuchipy.detectors._ebsd_detector import EBSDDetector
 from kikuchipy.io.plugins.nordif._api import _get_settings_from_file
